C# 11 goodies

+++ src/NetVips/Cache.cs
@@ -21,7 +21,7 @@ namespace NetVips
         /// </summary>
         public static ulong MaxMem
         {
-            get => (ulong) Vips.CacheGetMaxMem();
+            get => Vips.CacheGetMaxMem();
             set => Vips.CacheSetMaxMem(value);
         }
